pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cra-cli"
version = "0.1.0"
description = "EU Cyber Resilience Act compliance CLI — Product SBOM aggregation & VEX generation"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "CRA-CLI Contributors"},
]

dependencies = [
    "typer[all]>=0.9.0",
    "lib4sbom>=0.9.4",
    "lib4vex>=0.2.1",
    "csaf-tool>=0.3.2",
    "spdx-tools>=0.8.3",
    "markdown-it-py>=3.0.0",
    "packageurl-python>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
cra = "src.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
